refactor: extract DiversificationTypeButtons component
